Updated project to run on local, Refactored to cleanup, Added keyword find
RestaurantDAO can now find restaurants with certain keywords in the
restaurant info or its food items
