+++ pom.xml
@@ -14,7 +14,7 @@
     <name>RestaurantService</name>
     <description>Demo project for Spring Boot</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>1.8</java.version>
     </properties>
     <dependencies>
         <dependency>
